Add files via upload
These items in my system are all linked through the OPENGL, and include folders (instead of privately) so that it can be copied, and replicated to store a back up.  C:\SDK\OPENGL and C:\SDK\include  This is done through linker library and includes, if you are downloading the repository, then you should be able to simply move the OPENGL, and include folders to C:\SDK and build/run program.  Through OpenGL (GLFW) and ImGui this gives an example of making a shape appear through use of a checkbox.